--- robot_utils/robot_data/__init__.py
@@ -1,5 +1,0 @@
-from robot_utils.robot_data.array_data import ArrayData
-from robot_utils.robot_data.general_data import GeneralData
-from robot_utils.robot_data.img_data import ImgData
-from robot_utils.robot_data.pose_data import PoseData
-from robot_utils.robot_data.robot_data import RobotData
