+++ robotdatapy/data/__init__.py
@@ -1,0 +1,5 @@
+from robotdatapy.data.array_data import ArrayData
+from robotdatapy.data.general_data import GeneralData
+from robotdatapy.data.img_data import ImgData
+from robotdatapy.data.pose_data import PoseData
+from robotdatapy.data.robot_data import RobotData
