+++ setup.py
@@ -1,9 +1,9 @@
 from setuptools import setup, find_packages
 
 setup(
-    name='robot_utils',
+    name='robotdatapy',
     version='0.1.0',    
-    description='Robot Utilities Package',
+    description='Python package for interfacing with robot data',
     url='',
     author='Mason Peterson',
     author_email='masonbp@mit.edu',
